+++ index.js
@@ -14,14 +14,11 @@
 /*
  *  Components
  */
-export { register_c_yui_main, display_volatil_modal, display_info_message, display_warning_message, display_error_message, get_yesnocancel, get_yesno, get_ok } from "./src/c_yui_main.js";
 export { register_c_yui_window } from "./src/c_yui_window.js";
 export { register_c_yui_window_manager } from "./src/c_yui_window_manager.js";
-export { register_c_yui_tabs } from "./src/c_yui_tabs.js";
 export { register_c_yui_pager } from "./src/c_yui_pager.js";
 export { register_c_yui_wizard } from "./src/c_yui_wizard.js";
 export { register_c_yui_form } from "./src/c_yui_form.js";
-export { register_c_yui_routing } from "./src/c_yui_routing.js";
 export { register_c_yui_map } from "./src/c_yui_map.js";
 export { register_c_yui_uplot } from "./src/c_yui_uplot.js";
 export { register_c_yui_json_graph } from "./src/c_yui_json_graph.js";
@@ -70,8 +67,6 @@ export { register_c_g6_nodes_tree } from "./src/c_g6_nodes_tree.js";
 export { addClasses, removeClasses, toggleClasses, removeChildElements, disableElements, enableElements, set_submit_state, set_cancel_state, set_active_state, getStrokeColor } from "./src/lib_graph.js";
 export { inject_svg_icons } from "./src/lib_icons.js";
 export { EditControl, MarkerControl } from "./src/lib_maplibre.js";
-export { themes } from "./src/themes.js";
-export { YTable, createYTable } from "./src/ytable.js";
 export { yui_toolbar } from "./src/yui_toolbar.js";
 export { attach_clear } from "./src/yui_inputs.js";
 export { info_traffic, setup_dev, build_dev_panel, apply_dev_traces, dev_window_was_open } from "./src/yui_dev.js";
@@ -79,11 +74,8 @@ export { info_traffic, setup_dev, build_dev_panel, apply_dev_traces, dev_window_
 /*
  *  CSS - import these in your main entry point
  *  Example:
- *    import "gobj-ui/src/c_yui_main.css";
  *    import "gobj-ui/src/c_yui_shell.css";     // declarative shell
  *    import "gobj-ui/src/c_yui_map.css";
- *    import "gobj-ui/src/c_yui_routing.css";
- *    import "gobj-ui/src/ytable.css";
  *    import "gobj-ui/src/yui_toolbar.css";
  *    import "gobj-ui/src/lib_graph.css";
  *    import "gobj-ui/src/yui_icons.css";